Extract methods of Item class into separate traits
